[Issue 85] Move to Flink 1.4 - don't use deprecated class
